+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # master
 
+* We now require that all dependencies exist, and expect a `{ dependencies:
+  [...] }` object as return value from the callback
+
 # 1.0.1
 
 * Upgrade to node-merge-trees 2.0.0